+++ R/get_skimmers.R
@@ -1,41 +1,65 @@
-#' @include skim_with.R stats.R 
+#' @include skim_with.R stats.R
 NULL
 
 #' Retrieve the summary functions for a specific data type
-#' 
+#'
 #' These functions are used to set the default skimming functions for a data
 #' type.  When creating your own set of skimming functions, call [sfl()]
 #' within a [get_skimmers()] method for your particular type. Your call to
 #' [sfl()] should also provide a matching class in the `.type` argument.
 #' Otherwise, it will not be possible to dynamically reassign your default
 #' functions when working interactively.
-#' 
+#'
 #' Summary functions are provided for the following classes:
-#' 
+#'
 #'  - `numeric`
-#'  - `integer`
 #'  - `character`
 #'  - `factor`
 #'  - `logical`
 #'  - `complex`
 #'  - `Date`
 #'  - `POSIXct`
+#'  - `difftime`
 #'  - `ts`
 #'  - `AsIs`
 #'
 #' Call [get_default_skimmers()] to see the functions for each.
-#' 
+#'
 #' @param column An atomic vector or list. A column from a data frame.
 #' @return A `skim_function_list` object.
 #' @seealso [sfl()]
 #' @examples
 #' # Defining default skimming functions for a new class, `my_class`.
 #' # Note that the class argument is required for dynamic reassignment.
-#' get_skimmers.my_class <- function(colum) {
+#' get_skimmers.my_class <- function(column) {
 #'   sfl(
 #'     .type = "my_class",
 #'     mean,
-#'     sd)
+#'     sd
+#'   )
+#' }
+#'
+#' # Integer and float columns are both "numeric" and are treated the same
+#' # by default. To switch this behavior in another package, add a method.
+#' get_skimmers.integer <- function(column) {
+#'   sfl(
+#'     .type = "integer",
+#'     p50 = stats::quantile(
+#'       .,
+#'       probs = .50, na.rm = TRUE, names = FALSE, type = 1
+#'     )
+#'   )
+#' }
+#' x <- mtcars[c("gear", "carb")]
+#' class(x$carb) <- "integer"
+#' skim(x)
+#' \dontrun{
+#' # In a package, to revert to the V1 behavior of skimming separately with the
+#' # same functions, assign the numeric `get_skimmers`.
+#' get_skimmers.integer <- skimr::get_skimmers.numeric
+#'
+#' # Or, in a local session, use `skim_with` to create a different `skim`.
+#' new_skim <- skim_with(integer = skimr::get_skimmers.numeric())
 #' }
 #' @export
 get_skimmers <- function(column) {
@@ -56,19 +80,14 @@ get_skimmers.numeric <- function(column) {
     complete = n_complete,
     n = length,
     mean = mean(., na.rm = TRUE),
-    sd = sd(., na.rm = TRUE),
-    p0 = quantile(., probs = 0, na.rm = TRUE, names = FALSE),
-    p25 = quantile(., probs = .25, na.rm = TRUE, names = FALSE),
-    p50 = quantile(., probs= .50, na.rm = TRUE, names = FALSE),
-    p75 = quantile(., probs = .75, na.rm = TRUE, names = FALSE),
-    p100 = quantile(., probs = 1, na.rm = TRUE, names = FALSE),
-    hist = inline_hist(., 5))
-}
-
-#' @export
-get_skimmers.integer <- function(column) {
-  numeric_skimmers <- get_skimmers(numeric())
-  sfl(.type = "integer", !!!numeric_skimmers$keep)
+    sd = stats::sd(., na.rm = TRUE),
+    p0 = stats::quantile(., probs = 0, na.rm = TRUE, names = FALSE),
+    p25 = stats::quantile(., probs = .25, na.rm = TRUE, names = FALSE),
+    p50 = stats::quantile(., probs = .50, na.rm = TRUE, names = FALSE),
+    p75 = stats::quantile(., probs = .75, na.rm = TRUE, names = FALSE),
+    p100 = stats::quantile(., probs = 1, na.rm = TRUE, names = FALSE),
+    hist = inline_hist(., 5)
+  )
 }
 
 #' @export
@@ -80,20 +99,22 @@ get_skimmers.factor <- function(column) {
     n = length,
     ordered = is.ordered,
     n_unique = n_unique,
-    top_counts = top_counts)
+    top_counts = top_counts
+  )
 }
 
 #' @export
 get_skimmers.character <- function(column) {
   sfl(
     .type = "character",
-    missing  = n_missing,
+    missing = n_missing,
     complete = n_complete,
     n = length,
     min = min_char,
     max = max_char,
     empty = n_empty,
-    n_unique = n_unique)
+    n_unique = n_unique
+  )
 }
 
 #' @export
@@ -104,7 +125,8 @@ get_skimmers.logical <- function(column) {
     complete = n_complete,
     n = length,
     mean = mean(., na.rm = TRUE),
-    count = top_counts)
+    count = top_counts
+  )
 }
 
 #' @export
@@ -113,7 +135,8 @@ get_skimmers.complex <- function(column) {
     .type = "complex",
     missing = n_missing,
     complete = n_complete,
-    n = length)
+    n = length
+  )
 }
 
 #' @export
@@ -125,8 +148,9 @@ get_skimmers.Date <- function(column) {
     n = length,
     min = min(., na.rm = TRUE),
     max = max(., na.rm = TRUE),
-    median = median(., na.rm = TRUE),
-    n_unique = n_unique)
+    median = stats::median(., na.rm = TRUE),
+    n_unique = n_unique
+  )
 }
 
 #' @export
@@ -153,11 +177,12 @@ get_skimmers.ts <- function(column) {
     frequency = stats::frequency,
     deltat = stats::deltat,
     mean = mean(., na.rm = TRUE),
-    sd = sd(., na.rm = TRUE),
+    sd = stats::sd(., na.rm = TRUE),
     min = min(., na.rm = TRUE),
     max = max(., na.rm = TRUE),
-    median = median(., na.rm = TRUE),
-    line_graph  = inline_linegraph(., 16))
+    median = stats::median(., na.rm = TRUE),
+    line_graph = inline_linegraph(., 16)
+  )
 }
 
 #' @export
@@ -168,8 +193,9 @@ get_skimmers.list <- function(column) {
     complete = n_complete,
     n = length,
     n_unique = n_unique,
-    min_length= list_min_length,
-    max_length = list_max_length)
+    min_length = list_min_length,
+    max_length = list_max_length
+  )
 }
 
 #' @export
@@ -179,12 +205,13 @@ get_skimmers.AsIs <- function(column) {
 }
 
 #' @rdname get_skimmers
+#' @param class The class of the column being skimmed
 #' @export
 get_default_skimmers <- function(class = NULL) {
   if (is.null(class)) {
-    defaults <- as.character(methods("get_skimmers"))
+    defaults <- as.character(utils::methods("get_skimmers"))
     classes <- stringr::str_replace(defaults, "get_skimmers.", "")
-    no_default <- purrr::discard(classes, ~.x == "default")
+    no_default <- purrr::discard(classes, ~ .x == "default")
     iter <- purrr::set_names(no_default)
     purrr::map(iter, get_class_defaults)
   } else {