+++ R/skimr-package.R
@@ -4,12 +4,12 @@
 #' within R. The package's API is tidy, functions take data frames, return
 #' data frames and can work as part of a pipeline. The returned `skimr`
 #' object is subsettable and offers a human readable output.
-#' 
+#'
 #' `skimr` is opinionated, providing a strong set of summary statistics
 #' that are generated for a variety of different data types. It is also
 #' provides an API for customization. Users can change both the functions
 #' dispatched and the way the results are formatted.
-#' 
+#'
 #' @name skimr-package
 #' @aliases skimr
 #' @docType package