+++ R/stats.R
@@ -3,14 +3,14 @@
 #' Skimr provides extensions to a variety of functions with R's stats package
 #' to simplify creating summaries of data. All functions are vectorized and take
 #' a single argument. Other parameters for these functions are set in the
-#' [skim_format()] function.
+#' skim_format() function.
 #'
 #' @param x A vector
 #' @param n_bins In `inline_hist`, the number of histogram bars.
 #' @param length.out In `inline_linegraph`, the length of the character time
 #'   series.
 #' @param max_char In `top` = 3, max_levels = 4
-#' @seealso [skim_format()] and [purrr::partial()] for setting arguments of a
+#' @seealso [purrr::partial()] for setting arguments of a
 #'   skimmer function.
 #' @name stats
 NULL
@@ -40,10 +40,12 @@ n_complete <- function(x) {
 sorted_count <- function(x) {
   tab <- table(x, useNA = "always")
   names_tab <- names(tab)
-  if (is.element("",  names_tab)) {
+  if (is.element("", names_tab)) {
     names_tab[names_tab == ""] <- "empty"
-    warning("Variable contains value(s) of \"\" that have been ",
-            "converted to \"empty\".") 
+    warning(
+      "Variable contains value(s) of \"\" that have been ",
+      "converted to \"empty\"."
+    )
   }
   out <- rlang::set_names(as.integer(tab), names_tab)
   sort(out, decreasing = TRUE)
@@ -51,6 +53,7 @@ sorted_count <- function(x) {
 
 #' @describeIn stats Compute and collapse a contingency table into a single
 #'   character scalar. Wraps [sorted_count()].
+#' @param max_levels The maximum number of levels to be displayed.
 #' @export
 
 top_counts <- function(x, max_char = 3, max_levels = 4) {
@@ -72,16 +75,15 @@ top_counts <- function(x, max_char = 3, max_levels = 4) {
 inline_hist <- function(x, n_bins = 8) {
   # For the purposes of the histogram, treat infinite as NA
   # (before the test for all NA)
-  if (any(is.infinite(x))){
+  if (any(is.infinite(x))) {
     x[is.infinite(x)] <- NA
     warning(
       "Variable contains Inf or -Inf value(s) that were converted to NA."
-     )
+    )
   }
-  
+
   # Handle empty and NA vectors (is.na is TRUE for NaN)
-  if (length(x) < 1 || all(is.na(x)))
-  {
+  if (length(x) < 1 || all(is.na(x))) {
     return(" ")
   }
 
@@ -242,10 +244,10 @@ braille <- function(x) {
   x <- c(7L, 3L, 2L, 1L, 8L, 6L, 5L, 4L)[x]
 
   raised <- 1:8 %in% x
-  binary <- raised * 2 ^ (0:7)
+  binary <- raised * 2^(0:7)
 
   # offset in hex is 2800
-  val <- 10240 + sum(raised * 2 ^ (0:7))
+  val <- 10240 + sum(raised * 2^(0:7))
 
   intToUtf8(val)
 }
@@ -256,8 +258,11 @@ braille <- function(x) {
 list_lengths_min <- function(x) {
   x <- x[!is.na(x)]
   l <- lengths(x)
-  if (length(l) > 0) min(l)
-  else NA
+  if (length(l) > 0) {
+    min(l)
+  } else {
+    NA
+  }
 }
 
 
@@ -267,8 +272,11 @@ list_lengths_min <- function(x) {
 list_lengths_median <- function(x) {
   x <- x[!is.na(x)]
   l <- lengths(x)
-  if (length(l) > 0) stats::median(l)
-  else NA
+  if (length(l) > 0) {
+    stats::median(l)
+  } else {
+    NA
+  }
 }
 
 
@@ -278,15 +286,18 @@ list_lengths_median <- function(x) {
 list_lengths_max <- function(x) {
   x <- x[!is.na(x)]
   l <- lengths(x)
-  if (length(l) > 0) max(l)
-  else NA
+  if (length(l) > 0) {
+    max(l)
+  } else {
+    NA
+  }
 }
 
 
 #' @describeIn stats Get the length of the shortest list in a vector of lists.
 #' @export
 
-list_min_length <- function(x){
+list_min_length <- function(x) {
   l <- lengths(x)
   min(l)
 }
@@ -295,7 +306,7 @@ list_min_length <- function(x){
 #' @describeIn stats Get the length of the longest list in a vector of lists.
 #' @export
 
-list_max_length <- function(x){
+list_max_length <- function(x) {
   l <- lengths(x)
   max(l)
 }