+++ R/utils.R
@@ -24,10 +24,10 @@ is_windows <- function() {
 }
 
 #' Fix unicode histograms on Windows
-#' 
+#'
 #' This functions changes your session's locale to address issues with printing
 #' histograms on Windows.
-#' 
+#'
 #' There are known issues with printing the spark-histogram characters when
 #' printing a data frame, appearing like this: "<U+2582><U+2585><U+2587>".
 #' This longstanding problem originates in the low-level code for printing
@@ -35,10 +35,15 @@ is_windows <- function() {
 #' @export
 
 fix_windows_histograms <- function() {
-  message("This function will change your system locale. It may have other ",
-          "unintended effects.")
+  message(
+    "This function will change your system locale. It may have other ",
+    "unintended effects."
+  )
   response <- readline("Continue? (Y/n)")
-  if (tolower(response) != "n") Sys.setlocale("LC_CTYPE", "Chinese")
-  else message("Locale was not changed.")
+  if (tolower(response) != "n") {
+    Sys.setlocale("LC_CTYPE", "Chinese")
+  } else {
+    message("Locale was not changed.")
+  }
   invisible(NULL)
 }