+++ tests/testthat/test-sfl.R
@@ -8,14 +8,14 @@ test_that("sfl's require at least one function", {
   expect_error(sfl(.type = "numeric"))
 })
 
-test_that("The interface for sfl's separates keep and drop functions",{
+test_that("The interface for sfl's separates keep and drop functions", {
   input <- sfl(mad = mad, hist = NULL, .type = "test")
   expect_is(input, "skimr_function_list")
   expect_length(input, 3)
   expect_named(input, c("keep", "drop", "type"))
   expect_identical(input$drop, "hist")
   expect_identical(input$type, "test")
-  
+
   keep <- input$keep
   expect_is(keep, "fun_list")
   expect_true(attr(keep, "have_name"))