+++ NEWS.md
@@ -1,11 +1,21 @@
-skimr 1.0.3 (2018-xx-xx)
+skimr 1.0.4 (2018-01-05)
+========================
+### BUG FIXES
+  * Fix failures in handling dplyr verbs related to upcoming release
+     of dplyr 0.0.0.
+
+skimr 1.0.3 (2018-06-06)
 ========================
 ### NEW FEATURES
   * You can use skim_with() with a nest list of functions:
      `skim_with(.list = mylist)` or `skim_with(!!!mylist)`
+  * More polished display of subtables in default printing.
 
 ### BUG FIXES
-  * Fix issue with conflict between knitr and skimr versions of kable() that occurred intermittently.
+  * Fix issue with conflict between knitr and skimr versions 
+     of kable() that occurred intermittently.
+  * Do not skim a class when the skimmer list is empty for that class.
+  * Fix a mistake in a test of skim_print for top counts.
 
 skimr 1.0.2 (2018-04-04)
 ========================